+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.0.0
+
++ `BigIntJs`: JS implementation of `BigInt`
+
+### Breaking changes
+
++ `BigInt`: `BigInteger`
++ `BigIntVm`: `BigIntegerVm`
++ `BigIntRef`: `BigIntegerRef`
+
 ## 1.0.0
 
 + `BigInteger` interface desing